pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GovOn"
version = "1.0.1"
description = "On-Device AI Civil Complaint Analysis and Processing System using EXAONE-Deep-7.8B"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "On-Device AI Team", email = "team@example.com"}
]
keywords = [
    "llm",
    "fine-tuning",
    "quantization",
    "civil-complaint",
    "korean-nlp",
    "exaone",
    "awq",
    "qlora"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core ML/DL
    "torch>=2.8.0",
    "transformers>=4.53.0",
    "accelerate>=1.0.0",
    "datasets>=2.14.0",

    # Quantization & Optimization
    "bitsandbytes>=0.44.0",
    "optimum>=1.13.0",
    "autoawq>=0.2.8",

    # Fine-tuning
    "peft>=0.16.0",
    "trl>=0.15.0",

    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Utilities
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "bandit>=1.7.5",
    "safety>=2.3.5",
    "wandb>=0.16.0",
    "sacrebleu>=2.4.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
]

korean-nlp = [
    "konlpy>=0.6.0",
    "mecab-python3>=1.0.6",
]

inference = [
    "vllm>=0.14.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.7.4",
    "httpx>=0.27.0",
]

database = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
]

all = [
    "GovOn[dev,korean-nlp,inference,database]",
]

[project.urls]
Homepage = "https://github.com/GovOn-org/GovOn"
Documentation = "https://github.com/GovOn-org/GovOn#readme"
Repository = "https://github.com/GovOn-org/GovOn.git"
Issues = "https://github.com/GovOn-org/GovOn/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "src/inference/document_processor.py",
    "src/inference/db/database.py",
    "src/inference/db/alembic/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
